fixup!

+++ core/consensus/babe/types/babe_block_header.hpp
@@ -21,7 +21,7 @@ namespace kagome::consensus {
     /// output of VRF function
     crypto::VRFOutput vrf_output;
     /// authority index of the producer
-    size_t authority_index{};
+    uint32_t authority_index{};
   };
 
   /**
